refactor(cli): colocate command constructors
