--- go/core/cli/root_external_test.go
@@ -1,14 +1,0 @@
-package cli_test
-
-import (
-	"testing"
-
-	"github.com/kagent-dev/kagent/go/core/cli"
-)
-
-func TestRootIsExternallyImportable(t *testing.T) {
-	root := cli.Root()
-	if root == nil {
-		t.Fatal("Root() returned nil")
-	}
-}
